feat(llava_arch): (WIP) introduce LlavaMetaModel implementation
This is the main LLava model architecture. As done in the main LLaVa repo,
it will be composed of the following:
- a vision tower (here a CLIP model)
- a mm_projector (2MLP with Gelu)
it will have a main method  prepare_inputs_labels_for_multimodal that will convert
the optional image and the text prompt (after it is tokenized) to a flat, ready to
be fed to the LLM block, embeddings. The LLM block will be implemented later (LLaMa block).
The architecture and the weights that will be targeted are the one available
here: https://huggingface.co/llava-hf/llava-1.5-7b-hf